LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/redirected-response.https-expected.txt

CONSOLE MESSAGE: Not allowed to follow a redirection while loading https://localhost:9443/service-workers/service-worker/resources/redirect.py?Redirect=https%3A%2F%2Flocalhost%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Fsimple.txt%3F&error
CONSOLE MESSAGE: Fetch API cannot load https://localhost:9443/service-workers/service-worker/resources/redirect.py?Redirect=https%3A%2F%2Flocalhost%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Fsimple.txt%3F&error due to access control checks.
CONSOLE MESSAGE: Response served by service worker has redirections
CONSOLE MESSAGE: Fetch API cannot load https://localhost:9443/service-workers/service-worker/resources/simple.txt? due to access control checks.
CONSOLE MESSAGE: Response served by service worker has redirections
CONSOLE MESSAGE: Fetch API cannot load https://localhost:9443/service-workers/service-worker/resources/simple.txt? due to access control checks.
CONSOLE MESSAGE: Response served by service worker is opaque redirect
CONSOLE MESSAGE: Fetch API cannot load https://localhost:9443/service-workers/service-worker/resources/redirect.py?Redirect=https%3A%2F%2Flocalhost%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Fsimple.txt%3F&original-redirect-mode=follow&sw=manual due to access control checks.
CONSOLE MESSAGE: Response served by service worker is opaque redirect
CONSOLE MESSAGE: Fetch API cannot load https://localhost:9443/service-workers/service-worker/resources/redirect.py?Redirect=https%3A%2F%2Flocalhost%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Fsimple.txt%3F&original-redirect-mode=error&sw=manual due to access control checks.
CONSOLE MESSAGE: Not allowed to follow a redirection while loading https://localhost:9443/service-workers/service-worker/dummy?url=https%3A%2F%2Flocalhost%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Fsimple.txt%3F&original-redirect-mode=error&sw=gen
CONSOLE MESSAGE: Fetch API cannot load https://localhost:9443/service-workers/service-worker/dummy?url=https%3A%2F%2Flocalhost%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Fsimple.txt%3F&original-redirect-mode=error&sw=gen due to access control checks.

PASS initialize global state (service worker registration and caches) 
PASS mode: "follow", non-intercepted request, no server redirect 
PASS mode: "follow", non-intercepted request 
PASS mode: "manual", non-intercepted request 
PASS mode: "error", non-intercepted request 
PASS mode: "follow", no mode change, no server redirect 
PASS mode: "follow", no mode change 
PASS mode: "error", mode change: "follow" 
PASS mode: "manual", mode change: "follow" 
PASS mode: "follow", mode change: "manual" 
PASS mode: "error", mode change: "manual" 
PASS mode: "manual", no mode change 
PASS mode: "follow", generated redirect response 
PASS mode: "error", generated redirect response 
PASS mode: "manual", generated redirect response 
PASS Fetch should follow the redirect response 20 times 
PASS Fetch should not follow the redirect response 21 times. 
PASS The URL for the service worker redirected request should be propagated to response. 
PASS restore global state (service worker registration) 
PASS restore global state (caches) 
